requirement.txt
Django==4.2.0
psycopg2-binary==2.9.6
redis==5.0.0
django-redis==5.2.0
gunicorn==21.2.0
python-decouple==3.8
django-ninja==1.1.0
PyJWT==2.8.0
python-multipart==0.0.6